requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
